# api/openapi.yaml
openapi: 3.0.3
info:
  title: DispScenario Analyst API
  version: 0.1.0
  description: Contract for the complete v2 replacement.
servers:
  - url: http://localhost:8787
security:
  - BearerAuth: []
paths:
...
  /v1/analysis-runs/{runId}:
    get:
      operationId: getAnalysisRun
      parameters:
        - $ref: "#/components/parameters/RunId"
      responses:
        "200":
          description: Analysis run
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisRun"
  /v1/analysis-runs/{runId}/cancel:
    post:
      operationId: cancelAnalysisRun
      parameters:
        - $ref: "#/components/parameters/RunId"
      responses:
        "200":
          description: Cancelled run
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisRun"
  /v1/analysis-runs/{runId}/retry:
    post:
      operationId: retryAnalysisRun
      parameters:
        - $ref: "#/components/parameters/RunId"
      responses:
        "202":
          description: Retried run
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisRun"
...
  /v1/ground-truth/import:
    post:
      operationId: importGroundTruth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JsonObject"
      responses:
        "201":
          $ref: "#/components/responses/JsonObject"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    RunId:
      name: runId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    JsonObject:
      description: JSON object
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/JsonObject"
...
  schemas:
...
    JsonObject:
      type: object
      additionalProperties: true
...
    AnalysisRun:
      type: object
      required:
        - id
        - recordingId
        - status
        - provider
        - promptVersion
        - normalizationVersion
        - groupingVersion
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        recordingId:
          type: string
          format: uuid
        status:
          type: string
          enum: [QUEUED, PROCESSING, NORMALIZING, COMPLETED, FAILED, CANCELLED]
        provider:
          type: string
        model:
          type: string
          nullable: true
        promptVersion:
          type: string
        normalizationVersion:
          type: string
        groupingVersion:
          type: string
        error:
          type: string
          nullable: true
        inputTokens:
          type: integer
          format: int64
          minimum: 0
          nullable: true
        outputTokens:
          type: integer
          format: int64
          minimum: 0
          nullable: true
        thinkingTokens:
          type: integer
          format: int64
          minimum: 0
          nullable: true
        totalTokens:
          type: integer
          format: int64
          minimum: 0
          nullable: true
        estimatedCostUsd:
          type: number
          format: double
          minimum: 0
          nullable: true
        pricingVersion:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time